Added BIAS voltage parts

+++ src/control.c
@@ -46,7 +46,7 @@
 #include "fifo.h"
   
 
-double bias_voltage_range[NUM_BIAS_VOLTAGE] = {
+float bias_voltage_range[NUM_BIAS_VOLTAGE] = {
   9.0, -9.0, 9.0, -9.0, 9.0, -9.0, 9.0, -9.0, 9.0, -9.0, 
   9.0, -9.0, 99.0, 5.0, -15.0, -25.0, -10.0, -5.1, 0.0, 0.0
 };
